requirements.txt
requests
wxPython
pypubsub